ctr: move shim to commands
