+++ cmd/ctr/commands/shim/io_unix.go
@@ -1,6 +1,6 @@
 // +build !windows
 
-package main
+package shim
 
 import (
 	gocontext "context"
