+++ cmd/ctr/main_unix.go
@@ -2,6 +2,8 @@
 
 package main
 
+import "github.com/containerd/containerd/cmd/ctr/commands/shim"
+
 func init() {
-	extraCmds = append(extraCmds, shimCommand)
+	extraCmds = append(extraCmds, shim.Command)
 }
